+++ src/lpj/help.c
@@ -26,7 +26,7 @@
 #else
 #define IMAGE_OPT ""
 #endif
-#define USAGE "\nUsage: %s [-h] [-l] [-v] [-hash] [-vv] [-pedantic] [-ofiles] [-param] [-nopp] [-pp cmd] " FPE_OPT "\n" IMAGE_OPT\
+#define USAGE "\nUsage: %s [-h] [-l] [-v]  [-repo] [-hash] [-vv] [-pedantic] [-ofiles] [-param] [-nopp] [-pp cmd] " FPE_OPT "\n" IMAGE_OPT\
                 "       [-couple host[:port]] [-wait time]\n"\
                 "       [-outpath dir] [-inpath dir] [-restartpath dir]\n"\
                 "       [[-Dmacro[=value]] [-Idir] ...] filename\n"
@@ -57,6 +57,7 @@ void help(const char *progname /**< program filename */
           "-h,--help           print this help text\n"
           "-l,--license        print license file\n"
           "-v,--version        print version, compiler and compile flags\n"
+          "-repo               print GIT repository\n"
           "-hash               print GIT hash\n"
           "-vv                 verbosely print the actual values during reading of the\n"
           "                    configuration files\n"